Allow voice assist to function in a non-interactive state
Currently the InputDispatcher blocks the VOICE_ASSIST key from being
sent to the window manager when the system is in a non-interactive
state.  Move the key handling to interceptKeyBeforeQueueing to avoid
this but also push off the behavior to a handler to avoid any
potential blocking of the input queue as we call into other system
services.

Bug: 16292420
Change-Id: I5bbe3455f2af5249151127dede2204cf1f12a19f