Material design for screen pinning cling
Switching from a dialog to a cling-like view to be more material.
Also moving it into SysUI as it highlights the buttons in nav bar.
This will be triggered directly from recents, but from apps it
will be triggered by the PhoneStatusBar.  In the case that apps
trigger the request, there will be a 'no thanks' button so that
users can refuse.

Bug: 16957435
Change-Id: Ie880f82c66f4b2be98283ed3fa789703c4bd7ed5