Fix issue #18240452: No recents screenshot when resuming from keyguard.
WindowManager.keyguardGoingAway() isn't called early enough when you exit
the keyguard by clicking on a notification. So, the WindowState for the
preivous activity behind the keyguard is never transitioned to visible
and the activity manager then fails to take the screenshot for recents.
We will now be taking a screenshot of the activity before we go to sleep
so we are not dependent on signals from the keyguard.

Change-Id: I2acb2ad7a627d4e446ba11c9a0842d21fa6922d3